+++ src/main/java/com/thanhdat/quanlyhoctap/controller/NewsController.java
@@ -1,6 +1,5 @@
 package com.thanhdat.quanlyhoctap.controller;
 
-import com.thanhdat.quanlyhoctap.dto.request.CourseCrudRequest;
 import com.thanhdat.quanlyhoctap.dto.request.NewsCrudRequest;
 import com.thanhdat.quanlyhoctap.dto.response.*;
 import com.thanhdat.quanlyhoctap.service.NewsService;
@@ -25,7 +24,7 @@ public class NewsController {
     }
 
     @GetMapping("/view/{id}")
-    public ResponseEntity<NewsViewResponse> get(@PathVariable Integer id){
+    public ResponseEntity<NewsViewResponse> get(@PathVariable Long id){
         return ResponseEntity.ok(newsService.get(id));
     }
 
@@ -37,18 +36,18 @@ public class NewsController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
         newsService.delete(id);
         return ResponseEntity.ok().build();
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<NewsViewCrudResponse> getById(@PathVariable Integer id) {
+    public ResponseEntity<NewsViewCrudResponse> getById(@PathVariable Long id) {
         return ResponseEntity.ok(newsService.getById(id));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Void> update(@PathVariable Integer id, @RequestBody NewsCrudRequest updateRequest) {
+    public ResponseEntity<Void> update(@PathVariable Long id, @RequestBody NewsCrudRequest updateRequest) {
         newsService.update(id, updateRequest);
         return ResponseEntity.ok().build();
     }
