+++ src/main/java/com/thanhdat/quanlyhoctap/dto/request/EducationProgramCourseRequest.java
@@ -9,7 +9,7 @@ import lombok.experimental.FieldDefaults;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class EducationProgramCourseRequest {
-    Integer id;
+    Long id;
     Integer semester;
-    Integer courseId;
+    Long courseId;
 }
