+++ src/main/java/com/thanhdat/quanlyhoctap/repository/ExamRepository.java
@@ -8,9 +8,9 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface ExamRepository extends JpaRepository<Exam, Integer> {
+public interface ExamRepository extends JpaRepository<Exam, Long> {
     @Query("SELECT fx FROM Exam fx WHERE fx.courseClass.semester.id = :semesterId AND fx.type = :type")
-    List<Exam> findBySemesterIdAndType(Integer semesterId, ExamType type);
+    List<Exam> findBySemesterIdAndType(Long semesterId, ExamType type);
 
-    Optional<Exam> findByCourseClassIdAndType(Integer courseClassId, ExamType type);
+    Optional<Exam> findByCourseClassIdAndType(Long courseClassId, ExamType type);
 }
