+++ src/main/java/com/thanhdat/quanlyhoctap/repository/InvoiceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InvoiceRepository extends JpaRepository<Invoice, Integer> {
-    Invoice findByStudentIdAndSemesterId(Integer studentId, Integer semesterId);
+public interface InvoiceRepository extends JpaRepository<Invoice, Long> {
+    Invoice findByStudentIdAndSemesterId(Long studentId, Long semesterId);
 }
